requirements.txt
psycopg2-binary
sqlalchemy
sqlacodegen
python-dotenv
